# fix(via-router): use str::match_indices in path::split_into