--- a/crates/via-router/Cargo.toml
+++ b/crates/via-router/Cargo.toml
@@ -7,6 +7,3 @@ license = "MIT OR Apache-2.0"
 description = "A fast and flexible http router."
 homepage = "https://github.com/zacharygolba/via"
 repository = "https://github.com/zacharygolba/via"
-
-[dependencies]
-smallvec = "1.13"