--- a/crates/via-router/src/lib.rs
+++ b/crates/via-router/src/lib.rs
@@ -48,9 +48,13 @@ impl<T> Router<T> {
 
     pub fn visit(&self, path: &str) -> Vec<Result<Found, VisitError>> {
         let mut segments = Vec::with_capacity(8);
+        let mut results = Vec::with_capacity(8);
+        let nodes = &self.nodes;
 
         path::split(&mut segments, path);
-        visitor::visit(path, &self.nodes, &segments)
+        visitor::visit(&mut results, nodes, &segments, path);
+
+        results
     }
 }