--- a/crates/via-router/src/path/mod.rs
+++ b/crates/via-router/src/path/mod.rs
@@ -2,4 +2,4 @@ mod pattern;
 mod span;
 
 pub use pattern::{patterns, Param, Pattern};
-pub use span::{split_into, Span};
+pub use span::{split, Span};