--- a/crates/via-router/src/path/span.rs
+++ b/crates/via-router/src/path/span.rs
@@ -1,5 +1,12 @@
 use smallvec::SmallVec;
 
+/// Defines the behavior of a collection that can have elements appended
+/// to the end.
+///
+pub trait Push {
+    fn push(&mut self, span: Span);
+}
+
 /// A matched range in the url path.
 ///
 #[derive(Debug, PartialEq)]
@@ -8,111 +15,74 @@ pub struct Span {
     end: usize,
 }
 
-pub fn split(path: &'static str) -> Vec<Span> {
-    let mut segments = Vec::new();
-    let mut start = None;
+pub fn split_into(segments: &mut impl Push, path: &str) {
+    // Assume the byte position of the first span in path is `0`. Bounds checks
+    // are required before creating a Span with this value.
+    let mut start = 0;
 
-    for (index, byte) in path.bytes().enumerate() {
-        match (start, byte) {
-            // The start index is set and the current byte is a `/`. This is the
-            // end of the current segment.
-            (Some(from), b'/') => {
-                // Push the segment range to the segments vector.
-                segments.push(Span::new(from, index));
+    // The length of path could be the end offset of the last Span from path.
+    let len = path.len();
 
-                // Reset the start index.
-                start = None;
-            }
-
-            // The start index is set and the current byte is not a terminating
-            // `/`. Continue to the next byte.
-            (Some(_), _) => {}
-
-            // The start index is not set and the current byte is a `/`. Continue
-            // to the next byte.
-            (None, b'/') => {}
-
-            // The current byte is a `/` and the start index is not set. This is
-            // the start of a new segment.
-            (None, _) => {
-                // Set the start index to the current index.
-                start = Some(index);
-            }
+    // Iterate over the byte position of each occurrence of '/' in path.
+    for (end, _) in path.match_indices('/') {
+        // If a range exists with a length greater than `0` from start to end,
+        // append a Span to segments. This bounds check prevents an empty range
+        // from ending up in segments.
+        if end > start {
+            segments.push(Span::new(start, end));
         }
+
+        // Move start to the byte position after end.
+        start = end + 1;
     }
 
-    // It is likely that the last character in the path is not a `/`. Check if
-    // the start index is set. If so, push the last segment to the segments vec
-    // using the length of the path as the end index.
-    if let Some(from) = start {
-        segments.push(Span::new(from, path.len()));
+    // If a range exists with a length greater than `0` from start to the
+    // length of path, append a Span to segments. This condition is true
+    // when path does not end with `/`.
+    if len > start {
+        segments.push(Span::new(start, len));
     }
-
-    segments
 }
 
-#[inline]
-pub fn split_into(segments: &mut SmallVec<[Span; 5]>, path: &str) {
-    let mut start = None;
-
-    for (index, byte) in path.chars().enumerate() {
-        match (start, byte) {
-            // The start index is set and the current byte is a `/`. This is the
-            // end of the current segment.
-            (Some(from), '/') => {
-                // Push the segment range to the segments vector.
-                segments.push(Span::new(from, index));
-
-                // Reset the start index.
-                start = None;
-            }
-
-            // The start index is set and the current byte is not a terminating
-            // `/`. Continue to the next byte.
-            (Some(_), _) => {}
-
-            // The start index is not set and the current byte is a `/`. Continue
-            // to the next byte.
-            (None, '/') => {}
-
-            // The current byte is a `/` and the start index is not set. This is
-            // the start of a new segment.
-            (None, _) => {
-                // Set the start index to the current index.
-                start = Some(index);
-            }
-        }
+impl Push for SmallVec<[Span; 5]> {
+    #[inline]
+    fn push(&mut self, span: Span) {
+        SmallVec::push(self, span);
     }
+}
 
-    // It is likely that the last character in the path is not a `/`. Check if
-    // the start index is set. If so, push the last segment to the segments vec
-    // using the length of the path as the end index.
-    if let Some(from) = start {
-        segments.push(Span::new(from, path.len()));
+impl Push for Vec<Span> {
+    #[inline]
+    fn push(&mut self, span: Span) {
+        Vec::push(self, span);
     }
 }
 
 impl Span {
     /// Returns the start offset of the matched range.
     ///
+    #[inline]
     pub fn start(&self) -> usize {
         self.start
     }
 
     /// Returns the end offset of the matched range.
     ///
+    #[inline]
     pub fn end(&self) -> usize {
         self.end
     }
 }
 
 impl Span {
+    #[inline]
     pub(crate) fn new(start: usize, end: usize) -> Self {
         Self { start, end }
     }
 }
 
 impl Clone for Span {
+    #[inline]
     fn clone(&self) -> Self {
         Self {
             start: self.start,
@@ -125,9 +95,9 @@ impl Clone for Span {
 mod tests {
     use smallvec::SmallVec;
 
-    use super::{split, split_into, Span};
+    use super::{split_into, Span};
 
-    const PATHS: [&str; 15] = [
+    const PATHS: [&str; 16] = [
         "/home/about",
         "/products/item/123",
         "/blog/posts/2024/june",
@@ -143,9 +113,10 @@ mod tests {
         "/blog/posts/2024//june",
         "/user//profile/settings",
         "/services/contact//us",
+        "/",
     ];
 
-    fn get_expected_results() -> [Vec<Span>; 15] {
+    fn get_expected_results() -> [Vec<Span>; 16] {
         [
             vec![Span::new(1, 5), Span::new(6, 11)],
             vec![Span::new(1, 9), Span::new(10, 14), Span::new(15, 18)],
@@ -172,20 +143,8 @@ mod tests {
             ],
             vec![Span::new(1, 5), Span::new(7, 14), Span::new(15, 23)],
             vec![Span::new(1, 9), Span::new(10, 17), Span::new(19, 21)],
+            vec![],
         ]
-    }
-
-    #[test]
-    fn test_split() {
-        let expected_results = get_expected_results();
-
-        for (path_index, path_value) in PATHS.iter().enumerate() {
-            let segments = split(path_value);
-
-            for (segment_index, segment_value) in segments.into_iter().enumerate() {
-                assert_eq!(segment_value, expected_results[path_index][segment_index]);
-            }
-        }
     }
 
     #[test]
@@ -198,7 +157,11 @@ mod tests {
             split_into(&mut segments, path_value);
 
             for (segment_index, segment_value) in segments.into_iter().enumerate() {
-                assert_eq!(segment_value, expected_results[path_index][segment_index]);
+                assert_eq!(
+                    segment_value, expected_results[path_index][segment_index],
+                    "{} ({}, {:?})",
+                    path_value, segment_index, segment_value
+                );
             }
         }
     }