Simplifies the logic used to accumulate path segment ranges in via-router. Also refactors via-router to store path segment ranges in a `Vec` rather than a `SmallVec`. This changes brings a significant performance improvement to via-router and also reduces stack usage in `Router::visit`.

### Benchmarks

The benchmarks are apples to apples. The only difference is that I added additional test cases to see how the router performs as the complexity of the url path increases. 

*Before:*

```
test find_matches_simple       ... bench:         118.79 ns/iter (+/- 2.97)
test find_matches_nested_stack ... bench:         240.54 ns/iter (+/- 12.51)
test find_matches_nested_heap  ... bench:         289.47 ns/iter (+/- 12.16)
```

*After:*

```
test find_matches_2            ... bench:         117.23 ns/iter (+/- 3.17)
test find_matches_3            ... bench:         146.60 ns/iter (+/- 2.76)
test find_matches_4            ... bench:         194.86 ns/iter (+/- 2.38)
test find_matches_5            ... bench:         237.18 ns/iter (+/- 9.49)
test find_matches_6            ... bench:         233.41 ns/iter (+/- 6.69)
test find_matches_7            ... bench:         251.43 ns/iter (+/- 7.11)
```