@zacharygolba 2025-01-07
The changes in this PR are available via crates.io in version `2.0.0-beta.8`.